Rounding correctly

+++ PhysicalLedLine.py
@@ -53,9 +53,9 @@ class Ws2801LedLine(LedLine):
     else:
       for i in range(10):
         adjust = (i+1) / 10.0
-        self._pixels[idx] = (round(255 * color.get_red() * adjust),
-                             round(255 * color.get_green() * adjust),
-                             round(255 * color.get_blue()) * adjust)
+        self._pixels[idx] = (int(round(255 * color.get_red() * adjust)),
+                             int(round(255 * color.get_green() * adjust)),
+                             int(round(255 * color.get_blue()) * adjust))
         self._pixels.show()
 
   def PostUpdate(self):
